# H7 LED strip