--- a/src/link/stm32_flash_h750_1m.ld
+++ b/src/link/stm32_flash_h750_1m.ld
@@ -50,8 +50,6 @@ REGION_ALIAS("STACKRAM", DTCM_RAM)
 REGION_ALIAS("FASTRAM", DTCM_RAM)
 REGION_ALIAS("MAIN", FLASH)
 
-INCLUDE "stm32_h750_common.ld"
-
 SECTIONS
 {
   .DMA_RAM (NOLOAD) :
@@ -91,4 +89,5 @@ SECTIONS
   } >RAM
 }
 
+INCLUDE "stm32_h750_common.ld"
 INCLUDE "stm32_h750_common_post.ld"