--- a/src/main/drivers/light_ws2811strip.h
+++ b/src/main/drivers/light_ws2811strip.h
@@ -31,7 +31,7 @@
 #if defined(USE_WS2811_SINGLE_COLOUR)
 #define WS2811_DATA_BUFFER_SIZE    1
 #define WS2811_DMA_BUFFER_SIZE     (WS2811_DATA_BUFFER_SIZE * WS2811_BITS_PER_LED)
-// Do 2 extra iterations of the DMA transfer with the ouptut set to low to generate the > 50us delay.
+// Do 2 extra iterations of the DMA transfer with the output set to low to generate the > 50us delay.
 #define WS2811_DELAY_ITERATIONS    2
 #else
 #define WS2811_DATA_BUFFER_SIZE    WS2811_LED_STRIP_LENGTH
@@ -39,6 +39,23 @@
 #define WS2811_DELAY_BUFFER_LENGTH 42
 // number of bytes needed is #LEDs * 24 bytes + 42 trailing bytes)
 #define WS2811_DMA_BUFFER_SIZE     (WS2811_DATA_BUFFER_SIZE * WS2811_BITS_PER_LED + WS2811_DELAY_BUFFER_LENGTH)
+#endif
+
+#ifdef USE_LEDSTRIP_CACHE_MGMT
+// WS2811_DMA_BUFFER_SIZE is multiples of uint32_t
+// Number of bytes required for buffer
+#define WS2811_DMA_BUF_BYTES              (WS2811_DMA_BUFFER_SIZE * sizeof (uint32_t))
+// Number of bytes required to cache align buffer
+#define WS2811_DMA_BUF_CACHE_ALIGN_BYTES  ((WS2811_DMA_BUF_BYTES + 0x20) & ~0x1f)
+// Size of array to create a cache aligned buffer
+#define WS2811_DMA_BUF_CACHE_ALIGN_LENGTH (WS2811_DMA_BUF_CACHE_ALIGN_BYTES / sizeof (uint32_t))
+extern uint32_t ledStripDMABuffer[WS2811_DMA_BUF_CACHE_ALIGN_LENGTH];
+#else
+#if defined(STM32F1) || defined(STM32F3)
+extern uint8_t ledStripDMABuffer[WS2811_DMA_BUFFER_SIZE];
+#else
+extern uint32_t ledStripDMABuffer[WS2811_DMA_BUFFER_SIZE];
+#endif
 #endif
 
 #define WS2811_TIMER_MHZ           48
@@ -71,11 +88,6 @@ void setUsedLedCount(unsigned ledCount);
 
 bool isWS2811LedStripReady(void);
 
-#if defined(STM32F1) || defined(STM32F3)
-extern uint8_t ledStripDMABuffer[WS2811_DMA_BUFFER_SIZE];
-#else
-extern uint32_t ledStripDMABuffer[WS2811_DMA_BUFFER_SIZE];
-#endif
 extern volatile bool ws2811LedDataTransferInProgress;
 
 extern uint16_t BIT_COMPARE_1;