Shareable memory isn't working on the H7. Need to figure out why, but in the meantime, this is the same fix as used for DSHOT, using manual cache cleaning (flushing). Define USE_LEDSTRIP_CACHE_MGMT to use.