@jflyper 2020-12-01
I haven't been following the H7 cache coherency problem at all, but the LED strip for H7 has been working for 2 years. When and how the heck did it stop working???