@SteveCEvans 2020-12-01
After much digging I've got to the bottom of the cache coherency issues. The problems started when I move the .bss section out of RAM. Prior to doing that the DMA_RAM section was placed at `0x30000000`, but now, for example, it's at `0x3000fae0`. The `memProtConfigure()` routine adjusts the start of the region to align with the start of a cache line, but experimentation shows that (for the H7 at least) the region must be aligned to a multiple of its size, which must be a power of two.

As stated at https://developer.arm.com/documentation/dui0646/a/cortex-m7-peripherals/optional-memory-protection-unit/mpu-region-base-address-register?lang=en

> The base address is aligned to the size of the region. For example, a 64KB region must be aligned on a multiple of 64KB, for example, at 0x00010000 or 0x00020000.

A simple fix is to include `stm32_h750_common.ld` after the `DMA_RAM` and `DMA_RW_AXI` sections have been defined so they sits at the start of `D2_RAM`/`RAM` respectively.

An unfortunate impact of this is that unless we align both the start and end of the `DMA_RAM` section to the size, any other variables falling within that region will suffer a performance hit of using bufferable (write-through) cached memory rather than write back.

Both the `DMA_RAM` and `DMA_RW_AXI` sections are defined in a similar manner so this impact will affect memory immediately following both of these sections.

I've made changes to the linker scripts to address the alignment, but I recommend that `USE_DSHOT_CACHE_MGMT` and `USE_LEDSTRIP_CACHE_MGMT` be defined for H7 so that other memory is cached with write-back and doesn't suffer a performance hit.