@SteveCEvans 2020-12-02
I'll raise another PR for the `adcConversionBuffer` covered in `add_stm32H7xx.c`. There's a comment that states `// Cache coherency should be maintained by MPU facility` for data read in from the ADCs, but the `DMA_RAM` is configured to use cacheable write-through memory which won't handle data written by DMA without an invalidate. Indeed there's a comment in the definition of the `mpuRegions[]` array which reads `// Reading needs cache coherence operation`.

Similarly the `transponder` value defined in `transponder_ir_io_hal.c` needs proper cache management too.

And finally the buffers defined in `serial_uart.c`.

Should I raise a single PR to cover these, or individual ones?