@jflyper 2020-12-02
Oh wait. Cache invalidation is an expensive operation, isn't it?