@jflyper 2020-12-02
And why did you move bss out of RAM? Sounds weird.