@SteveCEvans 2020-12-02
> Oh wait. Cache invalidation is an expensive operation, isn't it?

Not really. For writes we’re writing fresh data into the buffer (or rather a copy in cache) and then cleaning the whole region to push it out to RAM. Alternatively with write through every access reaches RAM which is about 1/3 the speed of cache. That would affect more than just the intended buffer as explained above. I intend doing some timings for the different approaches.