@SteveCEvans 2020-12-02
> And why did you move bss out of RAM? Sounds weird.

Because with both SD card support and external FLASH enabled with USB MSC .bss overflowed RAM, so it’s been moved to D2 RAM. Ideally it should be in DTCM along with the .data section, but that causes the bootloader to hang for some reason.