@mikeller 2020-12-02
> Because with both SD card support and external FLASH enabled with USB MSC .bss overflowed RAM, so it’s been moved to D2 RAM. Ideally it should be in DTCM along with the .data section, but that causes the bootloader to hang for some reason.

I think this is now getting out of hand - we are changing multiple areas of the firmware to a suboptimal solution just because some manufacturer supplies their flight controllers with a broken bootloader.
My proposal is to revert all of this, move the DMA RAM to DTCM, and wait for the manufacturer to fix their bootloader.