@SteveCEvans 2020-12-02
We can’t DMA to/from DTCM on the H7 (for these DMA engines) unlike on the F7. So we need to handle the DMA cache coherency irrespective of if it’s in RAM or D2_RAM.

This PR has highlighted that our current cache coherency solution is broken/suboptimal and it addresses that.

Moving .bss into DTCM is desirable and has highlighted some issues, but is rather orthogonal to the cache coherency issues I’m fixing here.