@TheIsotopes 2020-12-02
@SteveCEvans Short question, 
does this fix only concern the H750 series from SPRacing or also FC's with H74x / H72x from other manufacturers?

Because, on my SPRacing H7Nano, the LED strip works without any problems.