@SteveCEvans 2020-12-02
@TheIsotopes Building with current master the DMA_RAM section for a H7NANO is:

```.DMA_RAM        0x000000003000d700     0x2e4c
                0x000000003000d700                . = ALIGN (0x20)
                0x000000003000d700                PROVIDE (dmaram_start = .)
                0x000000003000d700                _sdmaram = .
                0x000000003000d700                _dmaram_start__ = _sdmaram
 *(.DMA_RAM)
 .DMA_RAM       0x000000003000d700      0xb64 /var/folders/7k/11ftt0b56792cylqxw59w69m0000gp/T//betaflight_SPRACINGH7NANO.elf.DUlsf2.ltrans0.ltrans.o
                0x000000003000d700                adcConversionBuffer.lto_priv.0
                0x000000003000d70c                transponder
 .DMA_RAM       0x000000003000e264      0xca8 /var/folders/7k/11ftt0b56792cylqxw59w69m0000gp/T//betaflight_SPRACINGH7NANO.elf.DUlsf2.ltrans3.ltrans.o
                0x000000003000e264                ledStripDMABuffer
 .DMA_RAM       0x000000003000ef0c     0x1640 /var/folders/7k/11ftt0b56792cylqxw59w69m0000gp/T//betaflight_SPRACINGH7NANO.elf.DUlsf2.ltrans9.ltrans.o
                0x000000003001054c                PROVIDE (dmaram_end = .)
                0x000000003001054c                _edmaram = .
                0x000000003001054c                _dmaram_end__ = _edmaram
```

In `memProtConfigure()`, length is 0x2E4C, `flsl()` will return 14, making `msbpos` 13. `MPU_InitStruct.Size` would this be 0x2000. Unless I'm mistaken the call to `HAL_MPU_ConfigRegion()` would thus configure the region from 0x3000c000 to 0x3000e000 with the desired cache characteristics. `ledStripDMABuffer` falls outside this region so I'd expect you to have issues. Cache coherency is an odd thing though, sometimes it bites you straight away, sometimes not, which is why it can be painful to debug.