@SteveCEvans 2020-12-02
Should I fix the implementation of `memProtConfigure()` to cover twice the region size if the length isn't a power of two?

I believe the following:

```
            if (length == (1U << msbpos)) {
                msbpos += 1;
            }
```

Should be:

```
            if (length != (1U << msbpos)) {
                msbpos += 1;
            }
```