@jflyper 2020-12-03
So, write through is optimal and forced write back (cache invalidation?) is sub-optimal?