@jflyper 2020-12-03
What about the difference of write back/through speeds to AXIM SRAM and D2 SRAMs? Are they different?

I'm raising this question to justify the relocation of entire bss to D2 SRAM.