@SteveCEvans 2020-12-03
Some more timings. DSHOT timing, not directly related to this PR, but of relevance to the discussion on cache management...

Original timings using shareable memory.

- DSHOT telemetry input 6.5 us
- DSHOT output 4.1 us

![image](https://user-images.githubusercontent.com/11480839/101080105-44307d00-35a0-11eb-87cc-533d4943516c.png)

New timings using explicit cache management.

- DSHOT telemetry input 3.4 us
- DSHOT output 3.8 us

![image](https://user-images.githubusercontent.com/11480839/101080173-59a5a700-35a0-11eb-811f-c2ef7893af22.png)

Explicit cache management is the clear winner for DSHOT.